requirements.txt
avro-python3==1.9.1
confluent-kafka==1.3.0
#aiohttp==3.6.2
requests==2.9.1
numpy==1.18.2
guppy3==3.0.9
